chord: 계층형으로 전환
