+++ backend-server/src/main/java/com/viewpharm/yakal/annotation/DisableSwaggerSecurity.java
@@ -1,0 +1,11 @@
+package com.viewpharm.yakal.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface DisableSwaggerSecurity {
+}
