--- backend-server/src/main/java/com/viewpharm/yakal/common/advice/SuccessResponseAdvice.java
@@ -1,30 +1,0 @@
-package com.viewpharm.yakal.common.advice;
-
-import com.viewpharm.yakal.common.dto.ResponseDto;
-import org.springframework.core.MethodParameter;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import org.springframework.http.server.ServerHttpRequest;
-import org.springframework.http.server.ServerHttpResponse;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
-
-@RestControllerAdvice(basePackages = "com.viewpharm.yakal")
-public class SuccessResponseAdvice implements ResponseBodyAdvice {
-
-    @Override
-    public boolean supports(MethodParameter returnType, Class converterType) {
-        return true;
-    }
-
-    @Override
-    public Object beforeBodyWrite(Object body, MethodParameter returnType, MediaType selectedContentType,
-                                  Class selectedConverterType, ServerHttpRequest request, ServerHttpResponse response) {
-        if (returnType.getParameterType() == ResponseDto.class) {
-            HttpStatus status = ((ResponseDto<?>) body).getHttpStatus();
-            response.setStatusCode(status);
-        }
-
-        return body;
-    }
-}
