--- backend-server/src/main/java/com/viewpharm/yakal/common/annotation/DisableSwaggerSecurity.java
@@ -1,11 +1,0 @@
-package com.viewpharm.yakal.common.annotation;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.METHOD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface DisableSwaggerSecurity {
-}
