--- backend-server/src/main/java/com/viewpharm/yakal/common/jwt/dto/JwtTokenDto.java
@@ -1,13 +1,0 @@
-package com.viewpharm.yakal.common.jwt.dto;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.Getter;
-
-@Data
-@Getter
-@Builder
-public class JwtTokenDto {
-    private String accessToken;
-    private String refreshToken;
-}
