--- backend-server/src/main/java/com/viewpharm/yakal/common/jwt/resolver/UserIdArgumentResolver.java
@@ -1,25 +1,0 @@
-package com.viewpharm.yakal.common.jwt.resolver;
-
-import com.viewpharm.yakal.common.jwt.annotation.UserId;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Component
-public class UserIdArgumentResolver implements HandlerMethodArgumentResolver {
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterType().equals(String.class)
-                && parameter.hasParameterAnnotation(UserId.class);
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) throws Exception {
-        return Long.valueOf(((HttpServletRequest) webRequest).getAttribute("USER_ID").toString());
-    }
-}
