--- backend-server/src/main/java/com/viewpharm/yakal/common/validator/EnumValidator.java
@@ -1,32 +1,0 @@
-package com.viewpharm.yakal.common.validator;
-
-import com.viewpharm.yakal.common.annotation.Enum;
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-public class EnumValidator implements ConstraintValidator<Enum, java.lang.Enum> {
-
-    private Enum annotation;
-
-    @Override
-    public void initialize(Enum constraintAnnotation) {
-        this.annotation = constraintAnnotation;
-    }
-
-    @Override
-    public boolean isValid(java.lang.Enum value, ConstraintValidatorContext context) {
-        Object[] enumValues = this.annotation.enumClass().getEnumConstants();
-
-        if (enumValues == null) {
-            return false;
-        }
-
-        for (Object enumValue : enumValues) {
-            if (value == enumValue) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-}
