+++ backend-server/src/main/java/com/viewpharm/yakal/controller/AuthController.java
@@ -1,9 +1,9 @@
 package com.viewpharm.yakal.controller;
 
-import com.viewpharm.yakal.common.jwt.dto.JwtTokenDto;
-import com.viewpharm.yakal.common.jwt.service.JwtService;
+import com.viewpharm.yakal.dto.JwtTokenDto;
+import com.viewpharm.yakal.service.JwtService;
 import com.viewpharm.yakal.service.AuthService;
-import com.viewpharm.yakal.common.dto.ResponseDto;
+import com.viewpharm.yakal.dto.ResponseDto;
 import com.viewpharm.yakal.type.ELoginProvider;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
