+++ backend-server/src/main/java/com/viewpharm/yakal/exception/CommonException.java
@@ -1,0 +1,13 @@
+package com.viewpharm.yakal.exception;
+
+public class CommonException extends RuntimeException {
+    private final ErrorCode errorCode;
+
+    public CommonException(final ErrorCode errorCode) {
+        this.errorCode = errorCode;
+    }
+
+    public ErrorCode getErrorCode() {
+        return errorCode;
+    }
+}
