+++ backend-server/src/main/java/com/viewpharm/yakal/exception/GlobalExceptionHandler.java
@@ -1,0 +1,37 @@
+package com.viewpharm.yakal.exception;
+
+import com.viewpharm.yakal.dto.ResponseDto;
+import jakarta.validation.ConstraintViolationException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@Slf4j
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(value = {ConstraintViolationException.class})
+    public ResponseDto<?> handleConstraintViolationException(ConstraintViolationException e) {
+        log.error("handleConstraintViolationException() in GlobalExceptionHandler throw ConstraintViolationException : {}", e.getMessage());
+        return ResponseDto.toResponseEntity(e);
+    }
+
+    @ExceptionHandler(value = {MethodArgumentNotValidException.class})
+    public ResponseDto<?> handleArgumentNotValidException(MethodArgumentNotValidException e) {
+        log.error("handleArgumentNotValidException() in GlobalExceptionHandler throw MethodArgumentNotValidException : {}", e.getMessage());
+        return ResponseDto.toResponseEntity(e);
+    }
+
+    @ExceptionHandler(value = {CommonException.class})
+    public ResponseDto<?> handleApiException(CommonException e) {
+        log.error("handleApiException() in GlobalExceptionHandler throw CommonException : {}", e.getErrorCode());
+        return ResponseDto.toResponseEntity(e);
+    }
+
+    @ExceptionHandler(value = {Exception.class})
+    public ResponseDto<?> handleException(Exception e) {
+        log.error("handleException() in GlobalExceptionHandler throw Exception : {}", e.getMessage());
+        return ResponseDto.toResponseEntity(new CommonException(ErrorCode.SERVER_ERROR));
+    }
+}
