+++ backend-server/src/main/java/com/viewpharm/yakal/interceptor/ConvertJwtToUserIdInterceptor.java
@@ -1,0 +1,17 @@
+package com.viewpharm.yakal.interceptor;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+public class ConvertJwtToUserIdInterceptor implements HandlerInterceptor {
+
+    @Override
+    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        request.setAttribute("USER_ID", authentication.getName());
+        return HandlerInterceptor.super.preHandle(request, response, handler);
+    }
+}
