+++ backend-server/src/main/java/com/viewpharm/yakal/repository/UserRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, String> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findBySocialIdAndLoginProvider(String socialId, ELoginProvider loginProvider);
 
     @Query("SELECT u.id AS id, u.userRole AS userRole FROM User u WHERE u.id = :userId AND u.isLogin = true AND u.refreshToken = :refreshToken")
