+++ backend-server/src/main/java/com/viewpharm/yakal/security/JwtAccessDenied.java
@@ -1,0 +1,21 @@
+package com.viewpharm.yakal.security;
+
+import com.viewpharm.yakal.exception.ErrorCode;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.web.access.AccessDeniedHandler;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+public class JwtAccessDenied extends JwtErrorResponse implements AccessDeniedHandler {
+
+    @Override
+    public void handle(final HttpServletRequest request,
+                       final HttpServletResponse response,
+                       final AccessDeniedException accessDeniedException) throws IOException {
+        setErrorResponse(response, ErrorCode.ACCESS_DENIED_ERROR);
+    }
+}
