+++ backend-server/src/main/java/com/viewpharm/yakal/service/AuthService.java
@@ -4,10 +4,9 @@ import com.viewpharm.yakal.type.ELoginProvider;
 import com.viewpharm.yakal.domain.User;
 import com.viewpharm.yakal.repository.UserRepository;
 import com.viewpharm.yakal.type.EUserRole;
-import com.viewpharm.yakal.common.exception.CommonException;
-import com.viewpharm.yakal.common.exception.type.ErrorCode;
-import com.viewpharm.yakal.common.jwt.dto.JwtTokenDto;
-import com.viewpharm.yakal.common.jwt.service.JwtService;
+import com.viewpharm.yakal.exception.CommonException;
+import com.viewpharm.yakal.exception.ErrorCode;
+import com.viewpharm.yakal.dto.JwtTokenDto;
 import com.viewpharm.yakal.utils.OAuth2Util;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.transaction.Transactional;
