+++ backend-server/src/main/java/com/viewpharm/yakal/validator/EnumValidator.java
@@ -1,0 +1,32 @@
+package com.viewpharm.yakal.validator;
+
+import com.viewpharm.yakal.annotation.Enum;
+import jakarta.validation.ConstraintValidator;
+import jakarta.validation.ConstraintValidatorContext;
+
+public class EnumValidator implements ConstraintValidator<Enum, java.lang.Enum> {
+
+    private Enum annotation;
+
+    @Override
+    public void initialize(Enum constraintAnnotation) {
+        this.annotation = constraintAnnotation;
+    }
+
+    @Override
+    public boolean isValid(java.lang.Enum value, ConstraintValidatorContext context) {
+        Object[] enumValues = this.annotation.enumClass().getEnumConstants();
+
+        if (enumValues == null) {
+            return false;
+        }
+
+        for (Object enumValue : enumValues) {
+            if (value == enumValue) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+}
